Fix misleading throughput parameters
The throughput of the 100% write and 100% read benchmarks cannot be
configured.  Only the number of concurrent users can.  Those benchmark
are designed with a closed model [1] so that the number of requests
injection rate depends on the response rate, and so that the requests do
not queue up on the server side.

This commit renames the parameters and fixes the documentation
accordingly, for clarity.