Add TableCommitsCreator simulation
This new simulation adds a new workload that continuously updates the
properties of every table.  There is no bound to the maximum number of
properties that can be created this way.  Each new property results in a
new snapshot to be created for that particular table.  This can quickly
amount to hundreds of commits per table.

This simulation can serve as the basis for table metadata management
workloads.