[feat] Add commits creator simulation (#24)
This new simulation adds a new workload that continuously updates the
properties of every table and every view.  There is no bound to the
maximum number of properties that can be created this way.  Each new
property results in a new snapshot to be created for that particular
entity.  This can quickly amount to hundreds of commits entity.

This simulation can serve as the basis for table metadata management
workloads.