Cleanup Iceberg dependencies (#38)
Replace `iceberg.spark.runtime` dependency with the required individual Iceberg dependencies.

Initially `iceberg.spark.runtime` was used to speed up the development as no need to go through what dependencies are missing or needed for each catalog clients. Now that we have tests for most of the catalogs, it is easy to catch missing dependencies.